+++ src/components/common/MyModal/index.tsx
@@ -7,6 +7,7 @@ import {
   ModalContent,
   ModalFooter,
   ModalOverlay,
+  Portal,
 } from '@chakra-ui/react';
 import { ReactNode } from 'react';
 
@@ -29,37 +30,39 @@ const MyModal = ({
 }: MyModal) => {
   return (
     <>
-      <Modal isOpen={isOpen} onClose={onClose}>
-        <ModalOverlay />
-        <ModalContent maxW="modal.w" h="modal.h" borderRadius="10px">
-          <Heading
-            display="flex"
-            justifyContent="center"
-            alignItems="center"
-            fontSize="1.6rem"
-            bg="pink.300"
-            h="modal.header.h"
-            borderTopRadius="10px"
-            color="black"
-          >
-            {title}
-          </Heading>
-          <ModalCloseButton />
-          <ModalBody>{children}</ModalBody>
-          <ModalFooter justifyContent="center">
-            <Button
-              h="modal.button.h"
-              w="modal.button.w"
-              bg="pink.100"
-              onClick={onSubmit}
-              mb="28px"
-              color="pink.300"
+      <Portal>
+        <Modal isOpen={isOpen} onClose={onClose}>
+          <ModalOverlay />
+          <ModalContent maxW="modal.w" h="modal.h" borderRadius="10px">
+            <Heading
+              display="flex"
+              justifyContent="center"
+              alignItems="center"
+              fontSize="1.6rem"
+              bg="pink.300"
+              h="modal.header.h"
+              borderTopRadius="10px"
+              color="black"
             >
-              {buttonText}
-            </Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
+              {title}
+            </Heading>
+            <ModalCloseButton />
+            <ModalBody>{children}</ModalBody>
+            <ModalFooter justifyContent="center">
+              <Button
+                h="modal.button.h"
+                w="modal.button.w"
+                bg="pink.100"
+                onClick={onSubmit}
+                mb="28px"
+                color="pink.300"
+              >
+                {buttonText}
+              </Button>
+            </ModalFooter>
+          </ModalContent>
+        </Modal>
+      </Portal>
     </>
   );
 };
